added error handling to middleware

+++ server/index.js
@@ -380,6 +380,8 @@ app.delete('/api/post/:postId', (req, res, next) => {
   const { postId } = req.params;
   const value = [postId];
 
+  if (!Number(postId)) throw new ClientError(`${postId} must exist and has to be a positive integer`, 400);
+
   const sql = `
   DELETE FROM "posts"
   WHERE "postId" = $1
@@ -395,6 +397,8 @@ app.delete('/api/post/:postId', (req, res, next) => {
 app.delete('/api/profiles/:profileId', (req, res, next) => {
   const { profileId } = req.params;
   const value = [profileId];
+
+  if (!Number(profileId)) throw new ClientError(`${profileId} must exist and has to be a positive integer`, 400);
 
   const sql = `
   WITH "account.cte" AS (
@@ -417,6 +421,8 @@ app.delete('/api/accounts/:accountId', (req, res, next) => {
   const { accountId } = req.params;
   const value = [accountId];
 
+  if (!Number(accountId)) throw new ClientError(`${accountId} must exist and has to be a positive integer`, 400);
+
   const sql = `
   WITH "account.cte" AS (
   DELETE FROM "accounts"
@@ -436,6 +442,8 @@ app.delete('/api/accounts/:accountId', (req, res, next) => {
 app.delete('/api/user/:userId', (req, res, next) => {
   const { userId } = req.params;
   const value = [userId];
+
+  if (!Number(userId)) throw new ClientError(`${userId} must exist and has to be a positive integer`, 400);
 
   const sql = `
   DELETE FROM "users"
@@ -463,6 +471,15 @@ app.put('/api/profiles/:profileId', (req, res, next) => {
   returning *
   `;
 
+  if (!Number(profileId)) {
+    throw new ClientError(
+        `${profileId} must exist and has to be a positive integer`,
+        400
+    );
+  } else if (!name || !imgPath) {
+    throw new ClientError('Name and/or imgPath has to be defined', 400);
+  }
+
   db.query(sql, values)
     .then(result => {
       const data = result.rows;
@@ -476,6 +493,7 @@ app.put('/api/profiles/:profileId', (req, res, next) => {
     })
     .catch(err => next(err));
 });
+
 app.put('/api/user/username/:userId', (req, res, next) => {
   const { userId } = req.params;
   const { username } = req.body;
@@ -486,6 +504,15 @@ app.put('/api/user/username/:userId', (req, res, next) => {
   SET "username" = $1
   WHERE "userId" = $2
   `;
+
+  if (!Number(userId)) {
+    throw new ClientError(
+      `${userId} must exist and has to be a positive integer`,
+      400
+    );
+  } else if (!username) {
+    throw new ClientError('Username has to be defined', 400);
+  }
 
   db.query(sql, values)
     .then(result => {
@@ -514,6 +541,15 @@ app.put('/api/user/password/:userId', (req, res, next) => {
   SET "password" = $1
   WHERE "userId" = $2
   `;
+
+  if (!Number(userId)) {
+    throw new ClientError(
+        `${userId} must exist and has to be a positive integer`,
+        400
+    );
+  } else if (!password) {
+    throw new ClientError('password has to have a value', 400);
+  }
 
   db.query(sql, values)
     .then(result => {
